Discard helper decorator in favor of custom base helper
